requirements.txt
# Core Dependencies
langchain==0.3.15
langchain-community==0.3.15
langchain-openai==0.2.14

# Vector Store
faiss-cpu==1.13.2

# PDF Processing
pypdf==5.1.0

# LLM and Embeddings
openai==1.59.5
sentence-transformers==3.3.1
transformers==4.48.1
torch==2.6.0

# HuggingFace (optional, for free LLM)
huggingface-hub==0.27.1

# Frontend
streamlit==1.41.1

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1

# Additional utilities
tiktoken==0.8.0
numpy==2.2.2

# Optional: For creating sample PDFs
reportlab==4.0.7
